Change all test headers to source files with related modifications
Issue #299
