+++ tests/cell/cell_test_suite.cpp
@@ -25,9 +25,9 @@
 
 #include <sstream>
 
-#include <helpers/test_suite.hpp>
-#include <helpers/assertions.hpp>
 #include <xlnt/xlnt.hpp>
+#include <helpers/assertions.hpp>
+#include <helpers/test_suite.hpp>
 
 class cell_test_suite : public test_suite
 {
@@ -133,14 +133,13 @@ private:
         xlnt::workbook wb;
 
         const auto datatypes =
-        {
-            xlnt::cell::type::empty,
-            xlnt::cell::type::boolean,
-            xlnt::cell::type::error,
-            xlnt::cell::type::formula_string,
-            xlnt::cell::type::number,
-            xlnt::cell::type::shared_string
-        };
+            {
+                xlnt::cell::type::empty,
+                xlnt::cell::type::boolean,
+                xlnt::cell::type::error,
+                xlnt::cell::type::formula_string,
+                xlnt::cell::type::number,
+                xlnt::cell::type::shared_string};
 
         for (const auto &datatype : datatypes)
         {
@@ -208,7 +207,6 @@ private:
         xlnt_assert(!cell.has_formula());
     }
 
-
     void test_not_formula()
     {
         xlnt::workbook wb;
@@ -227,7 +225,7 @@ private:
         auto ws = wb.active_sheet();
         auto cell = ws.cell(xlnt::cell_reference(1, 1));
 
-        for (auto value : { true, false })
+        for (auto value : {true, false})
         {
             cell.value(value);
             xlnt_assert(cell.data_type() == xlnt::cell::type::boolean);
@@ -330,8 +328,8 @@ private:
         auto cell = ws.cell(xlnt::cell_reference(1, 1));
 
         // The bytes 0x00 through 0x1F inclusive must be manually escaped in values.
-        auto illegal_chrs = { 0, 1, 2, 3, 4, 5, 6, 7, 8, 11, 12, 14, 15, 16, 17, 18,
-                 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31 };
+        auto illegal_chrs = {0, 1, 2, 3, 4, 5, 6, 7, 8, 11, 12, 14, 15, 16, 17, 18,
+            19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31};
 
         for (auto i : illegal_chrs)
         {
@@ -340,9 +338,9 @@ private:
         }
 
         cell.value(std::string(1, 33));
-        cell.value(std::string(1, 9));  // Tab
-        cell.value(std::string(1, 10));  // Newline
-        cell.value(std::string(1, 13));  // Carriage return
+        cell.value(std::string(1, 9)); // Tab
+        cell.value(std::string(1, 10)); // Newline
+        cell.value(std::string(1, 13)); // Carriage return
         cell.value(" Leading and trailing spaces are legal ");
     }
 
@@ -392,8 +390,8 @@ private:
         auto cell = ws.cell("A1");
 
         xlnt::fill fill(xlnt::pattern_fill()
-            .type(xlnt::pattern_fill_type::solid)
-            .foreground(xlnt::color::red()));
+                            .type(xlnt::pattern_fill_type::solid)
+                            .foreground(xlnt::color::red()));
         cell.fill(fill);
 
         xlnt_assert(cell.has_format());
@@ -677,3 +675,5 @@ private:
         xlnt_assert_throws(cell.comment(), xlnt::exception);
     }
 };
+
+static cell_test_suite x{};
