+++ tests/cell/rich_text_test_suite.cpp
@@ -27,8 +27,8 @@
 #include <iostream>
 #include <sstream>
 
-#include <helpers/test_suite.hpp>
 #include <xlnt/xlnt.hpp>
+#include <helpers/test_suite.hpp>
 
 class rich_text_test_suite : public test_suite
 {
@@ -102,3 +102,4 @@ public:
         xlnt_assert_differs(text_formatted, text_family_differs);
     }
 };
+static rich_text_test_suite x{};
