+++ tests/utils/helper_test_suite.cpp
@@ -30,6 +30,8 @@
 #include <helpers/xml_helper.hpp>
 #include <helpers/assertions.hpp>
 
+#include <xlnt/utils/exceptions.hpp>
+
 class helper_test_suite : public test_suite
 {
 public:
@@ -154,3 +156,4 @@ public:
         xlnt_assert(xml_helper::compare_xml_exact("<a>text</a>", "<a>text</a>", true));
     }
 };
+static helper_test_suite x;
