+++ tests/workbook/named_range_test_suite.cpp
@@ -257,3 +257,4 @@ public:
             [range_as_string(range, include_value = True) for range in wbcopy.get_named_ranges()])*/
     }
 };
+static named_range_test_suite x;
