+++ tests/worksheet/page_setup_test_suite.cpp
@@ -25,30 +25,40 @@
 
 #include <iostream>
 
-#include <helpers/path_helper.hpp>
-#include <helpers/temporary_file.hpp>
 #include <helpers/test_suite.hpp>
 #include <xlnt/xlnt.hpp>
 
-class path_test_suite : public test_suite
+class page_setup_test_suite : public test_suite
 {
 public:
-    path_test_suite()
+    page_setup_test_suite()
     {
-        register_test(test_exists);
+        register_test(test_properties);
     }
 
-	void test_exists()
-	{
-		temporary_file temp;
+    void test_properties()
+    {
+        xlnt::page_setup ps;
 
-		if (temp.get_path().exists())
-		{
-			path_helper::delete_file(temp.get_path());
-		}
+        xlnt_assert_equals(ps.paper_size(), xlnt::paper_size::letter);
+        ps.paper_size(xlnt::paper_size::executive);
+        xlnt_assert_equals(ps.paper_size(), xlnt::paper_size::executive);
 
-		xlnt_assert(!temp.get_path().exists());
-		std::ofstream stream(temp.get_path().string());
-		xlnt_assert(temp.get_path().exists());
-	}
+        xlnt_assert(!ps.orientation_.is_set());
+        ps.orientation_.set(xlnt::orientation::landscape);
+        xlnt_assert_equals(ps.orientation_.get(), xlnt::orientation::landscape);
+
+        xlnt_assert(!ps.fit_to_page());
+        ps.fit_to_page(true);
+        xlnt_assert(ps.fit_to_page());
+
+        xlnt_assert(!ps.fit_to_height());
+        ps.fit_to_height(true);
+        xlnt_assert(ps.fit_to_height());
+
+        xlnt_assert(!ps.fit_to_width());
+        ps.fit_to_width(true);
+        xlnt_assert(ps.fit_to_width());
+    }
 };
+static page_setup_test_suite x;
